Refactor project structure into backend and frontend directories
