+++ README.md
@@ -16,9 +16,12 @@ Trong hướng dẫn này, chúng ta sẽ xây dựng và triển khai một age
 
 ```text
 support-agent/
+├── backend/
+│   ├── app.py
+│   └── customer_support.py
+├── frontend/
+├── tests/
 ├── .env
-├── customer_support.py
-├── app.py
 ├── requirements.txt
 ├── Dockerfile        # tùy chọn
 └── README.md         # bạn đang xem ở đây
@@ -72,10 +75,10 @@ Trong ví dụ này, biến `GEMINI_MODEL` được đặt mặc định là `ge
 
 > **Lưu ý:** Bạn cũng có thể đặt biến `GOOGLE_API_KEY`, gói `langchain-google-genai` sẽ tự động lấy giá trị này.
 
-## 5. Định nghĩa Workflow (`customer_support.py`)
+## 5. Định nghĩa Workflow (`backend/customer_support.py`)
 
 ```python
-# customer_support.py
+# backend/customer_support.py
 from typing import TypedDict
 from langgraph.graph import StateGraph, END
 from langchain_core.prompts import ChatPromptTemplate
@@ -195,10 +198,10 @@ def run_customer_support(query: str) -> dict:
     return {k: res[k] for k in ("category", "sentiment", "response")}
 ```
 
-## 6. Xây dựng Server API (`app.py`)
+## 6. Xây dựng Server API (`backend/app.py`)
 
 ```python
-# app.py
+# backend/app.py
 from fastapi import FastAPI
 from pydantic import BaseModel
 from dotenv import load_dotenv
@@ -229,7 +232,7 @@ def health_check():
 ## 7. Chạy trên máy cục bộ
 
 ```bash
-uvicorn app:app --host 0.0.0.0 --port 8000 --reload
+uvicorn backend.app:app --host 0.0.0.0 --port 8000 --reload
 ```
 
 * Truy cập [http://localhost:8000/docs](http://localhost:8000/docs) để xem giao diện tương tác của API.
@@ -254,7 +257,7 @@ RUN pip install --no-cache-dir -r requirements.txt
 
 COPY . .
 EXPOSE 8000
-CMD ["uvicorn", "app:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "backend.app:app", "--host", "0.0.0.0", "--port", "8000"]
 ```
 
 ```bash
